added function to save frames

+++ src/utils/utils.cpp
@@ -5,7 +5,7 @@ namespace framegen::utils {
 std::vector<float> make_times(int nFrames) {
     std::vector<float> times;
     times.reserve(nFrames);
-    float step = 1.0 / nFrames + 1;
+    float step = 1.0f / static_cast<float>(nFrames + 1);
     float cur = step;
     for (int i = 0; i < nFrames; i++) {
         times.push_back(cur);
